requirements.txt
streamlit==1.29.0
tensorflow==2.14.0
numpy==1.25.0
pillow==10.0.0
gdown==4.7.1
